% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{collect}
\alias{collect}
\alias{collect.aaedb_promise}
\title{Execute an `aaedb_promise` query and return the requested
  data table}
\usage{
\method{collect}{aaedb_promise}(x, ..., n = Inf, warn_incomplete = TRUE, cte = FALSE)
}
\arguments{
\item{x}{an \code{aaedb_promise} object generated with \link[aae.db]{fetch_table},
\link[aae.db]{fetch_query}, or \link[aae.db]{fetch_project}}

\item{\dots}{ignored}

\item{n}{argument passed to \link[dplyr]{collect} (defaults to \code{Inf})}

\item{warn_incomplete}{argument passed to \link[dplyr]{collect}
(defaults to \code{TRUE})}

\item{cte}{argument passed to \link[dplyr]{collect}
(defaults to \code{FALSE})}
}
\description{
\code{collect} executes a query and returns the requested
  data. Queries can be modified prior to collecting, for example,
  using \code{dplyr} tools such as \link[dplyr]{filter} and
  \link[dplyr]{select}
}
\examples{
# connect to the AAEDB
aaedb_connect()

# download the VEFMAP database flat file
vefmap <- fetch_table("v_vefmap_only_flat_data")
vefmap <- vefmap \%>\% collect()

# use dplyr operations to reduce a query prior to downloading data
ovens_sites <- fetch_table("site")
ovens_sites <- ovens_sites \%>\%
  filter(waterbody == "Ovens River") \%>\%
  collect()

# optional: disconnect from the AAEDB prior to ending the R session
# aaedb_disconnect()

}
