% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{fetch_data}
\alias{fetch_data}
\alias{fetch_table}
\alias{fetch_project}
\alias{fetch_cpue}
\alias{fetch_birds}
\title{Query the AAEDB and return a query object to be manipulated
  and evaluated}
\usage{
fetch_table(x, schema = "aquatic_data", collect = FALSE, ...)

fetch_project(project_id = NULL, reporting_system = NULL, collect = FALSE, ...)

fetch_cpue(
  project_id = NULL,
  reporting_system = NULL,
  collect = FALSE,
  criterion = NULL,
  ...
)

fetch_birds(icon_site = NULL, type = NULL, collect = FALSE, ...)
}
\arguments{
\item{x}{a character specifying an individual table in the AAEDB}

\item{schema}{schema in which \code{x} is found. Defaults to
\code{"aquatic_data"}}

\item{collect}{logical: should a query be executed (\code{TRUE}) or
evaluated lazily (\code{FALSE}, the default)}

\item{\dots}{additional arguments passed to \link[dbplyr]{tbl_sql}}

\item{project_id}{an integer specifying an individual AAE project.
Any integer values will be accepted but non-existent projects will
return an empty query. Use \code{list_projects} to see a list of all
current projects}

\item{reporting_system}{a character specifying an individual reporting
system, which may encompass multiple waterbodies. Reporting systems
are specified for only a subset of waterbodies; will return an empty
table if a reporting system is not specified. Use \code{list_systems}
to see a list of all current systems}

\item{criterion}{\code{list} containing `var`, `lower`, and `upper` elements
specifying bounds on `var`. Allows complex filtering such as restricting
CPUE estimates to a specific length or weight range}

\item{icon_site}{the TLM Icon Site for which data are required. Applicable
only to TLM bird data sets queried with \code{fetch_birds}}

\item{type}{the TLM bird survey type. One of \code{woodland birds} or
\code{waterbirds}. Applicable only to TLM bird data sets queried with
\code{fetch_birds}}
}
\description{
\code{fetch_table} and \code{fetch_project}
  represent two ways to interact with the AAEDB.
  \code{fetch_table} provides access to prepared tables in
  the database and \code{fetch_project} selects data for an
  individual AAE project.

  \code{fetch_cpue} and \code{fetch_birds} are wrappers to calculate full
  CPUE data sets for fish or full count data sets for birds, respectively.
  These functions create full survey tables, which accounts for zero
  catch/count records in surveyed sites where a particular species was
  not recorded. These records are excluded from queries to
  \code{fetch_project}.

  All functions require credentials to access the AAEDB, as well
  as a appropriate VPN connection. If making multiple queries,
  it can be easier to connect once to the AAEDB rather than repeatedly
  connecting (and disconnecting). This is possible with the
  \link[aae.db]{aaedb_connect} function.

  Update for version 0.1.0: the \code{fetch_} functions now return an
  unevaluated query rather than a full data table. This allows further
  filtering or changes to the query using \code{dplyr} methods prior to
  downloading the data. This is especially useful when downloading a subset
  of a much larger data sets. All three methods will remain available in
  future versions but may be renamed to better reflect their intended use.
}
\examples{
# connect to the AAEDB
aaedb_connect()

# dplyr methods used below
library(dplyr)

# set up a query that includes the full VEFMAP data set
vefmap <- fetch_project(2)

# can manipulate and filter this query with dplyr methods
vefmap <- vefmap |>
  filter(
    waterbody == "Campaspe River",
    scientific_name == "Maccullochella peelii"
  )

# evaluate this query with collect
vefmap <- vefmap |> collect()

# process a nested query using raw tables from the database
#   and `dplyr` methods
site_data <- fetch_table("site")
survey_data <- fetch_table("survey")
survey_info_dplyr <- site_data |>
  left_join(
    survey_data |> distinct(id_site, id_project),
    by = join_by(id_site)
  ) |>
  filter(grepl("ovens", waterbody, ignore.case = TRUE)) |>
  distinct(waterbody, id_project) |>
  arrange(waterbody, id_project) |>
  collect()

# and grab information for individual projects
ovens_data <- fetch_project(9)
ovens_data <- ovens_data |> collect()

# subset this to 2015-2017 surveys
ovens_data <- fetch_project(9)
ovens_data <- ovens_data |>
  filter(survey_year \%in\% c(2015:2017)) |>
  collect()

# download TLM bird data for the Barmah Icon Site woodland birds
birds <- fetch_birds(
  icon_site = "Barmah System",
  type = "woodland birds"
) |>
  collect()

# and download species or habitat info associated with these surveys
species_info <- birds |> fetch_species_info()
habitat <- birds |> fetch_habitat_info()

# optional: disconnect from the AAEDB prior to ending the R session
#   when all queries and evaluation is complete
# aaedb_disconnect()

}
