% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_info.R
\name{fetch_info}
\alias{fetch_info}
\alias{fetch_site_info}
\alias{fetch_survey_info}
\alias{fetch_species_info}
\title{Extract information on sites or species represented in the AAEDB}
\usage{
fetch_site_info(x = NULL, ..., collect = FALSE)

fetch_survey_info(x = NULL, ..., collect = FALSE)

fetch_species_info(x = NULL, ..., collect = FALSE)
}
\arguments{
\item{x}{an evaluated or unevaluated query (table) for which to extract
site or species information with \code{fetch_site_info} and
\code{fetch_species_info}}

\item{\dots}{additional arguments used to filter outputs
based on regex expressions (see \link[base]{grepl} for possible
arguments)}

\item{collect}{logical: should a query be executed (\code{TRUE}) or
evaluated lazily (\code{FALSE}, the default)}
}
\description{
\code{fetch_site_info}, \code{fetch_survey_info}, and
  \code{fetch_species_info} return details of sites, surveys, or species
  included in the AAEDB. This information can
  be filtered to sites or species represented in a query or data set or
  based on regex expressions. See examples for usage.
}
\examples{
# connect to the AAEDB
aaedb_connect()

# dplyr methods used below
library(dplyr)

# set up a query that includes the full VEFMAP data set
vefmap <- fetch_project(2)

# can manipulate and filter this query with dplyr methods
vefmap <- vefmap \%>\%
  filter(
    waterbody == "Campaspe River",
    scientific_name == "Maccullochella peelii"
  )

# evaluate this query with collect
vefmap <- vefmap \%>\% collect()

# fetch information on the sites in a data set
vefmap_site_info <- fetch_site_info(vefmap)

# "spatialise" this information with the `sf` package
library(sf)
vefmap_site_info <- vefmap_site_info \%>\%
  filter(!is.na(geom_pnt)) \%>\%
  collect()
vefmap_sf <- vefmap_site_info \%>\%
  st_set_geometry(st_as_sfc(vefmap_site_info$geom_pnt))

# and make a basic map with the `mapview` package
library(mapview)
vefmap_sf \%>\%
  select(-geom_pnt) \%>\%
  mapview(
    col.regions = "DarkGreen",
    label = "site_name",
    layer.name = "Survey site"
  )

# alternatively, extract information on all sites where any field of the
#   site table matches a regex pattern
murray_site_info <- fetch_site_info(pattern = "^Murray", ignore.case = TRUE)

# fetch information on the surveys in a data set
vefmap_survey_info <- fetch_survey_info(vefmap)

# extract information on the species in a data set
fetch_species_info(vefmap)

# or for all species with scientific names matching a regex pattern
fetch_species_info(pattern = "Maccull", ignore.case = TRUE)

# list all taxonomic groups for use in fetch_species_info
fetch_table("taxon_lu", collect = FALSE) \%>\%
  select(taxon_type) \%>\%
  collect() \%>\%
  pull(taxon_type) \%>\%
  unique()

# or list all primary disciplines for use in fetch_species_info
fetch_table("taxon_lu", collect = FALSE) \%>\%
  select(primary_discipline) \%>\%
  collect() \%>\%
  pull(primary_discipline) \%>\%
  unique()

# and download data for one group (setting primary_discipline to NULL
#   to override default setting)
fetch_species_info(pattern = "Aquatic invertebrates", ignore.case = TRUE)

# optional: disconnect from the AAEDB prior to ending the R session
#   when all queries and evaluation is complete
# aaedb_disconnect()

}
