% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tables.R
\name{list_tables}
\alias{list_tables}
\alias{list_projects}
\alias{list_systems}
\title{List available tables in the AAEDB}
\usage{
list_tables(schema = "aquatic_data", ..., collect = FALSE)

list_projects(..., collect = FALSE)

list_systems(..., collect = FALSE)
}
\arguments{
\item{schema}{the schema from which to list tables, defaults to
\code{"aquatic_data"} and can be one of \code{"aquatic_data"},
\code{"public"}, \code{"spatial"}, \code{"spatial_isc"},
or \code{"stream_network"},}

\item{\dots}{additional arguments passed to \link[dbplyr]{tbl_sql}}

\item{collect}{logical: should a query be executed (\code{TRUE}) or
evaluated lazily (\code{FALSE}, the default)}
}
\description{
\code{list_tables} returns a list of all tables in a
  specified schema in the AAEDB. This information can be used to
  view and generate queries based on all tables in the database. See
  \code{fetch_data} for examples of generating queries.
}
\examples{
# connect to the AAEDB
aaedb_connect()

# dplyr methods used below
library(dplyr)

# list all tables in the aquatic_data schema
list_tables() |> collect()

# list all tables in the spatial schema
list_tables(schema = "spatial") |> collect()

# optional: disconnect from the AAEDB prior to ending the R session
#   when all queries and evaluation is complete
# aaedb_disconnect()

}
