% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaedb_connect.R
\name{aaedb_connect}
\alias{aaedb_connect}
\title{Open a connection to the AAEDB}
\usage{
aaedb_connect(...)
}
\arguments{
\item{\dots}{ignored}
}
\description{
\code{aaedb_connect} opens a connection to the AAEDB,
  which can be useful when making repeated queries to the AAEDB as
  it avoids repeatedly entering user credentials. Connections to the
  AAEDB require an appropriate VPN connection.

  An open connection will be automatically detected by the
  \code{fetch_table} and \code{fetch_query} functions.

  It is good practice to close the database connection with
  \code{dbDisconnect(con)} when all queries are completed. This
  will happen automatically when the current R session ends. If
  the connection needs to be closed prior to this, it can be
  manually closed by running \code{dbDisconnect(DB_ENV$conn)}.
}
\examples{
# connect to the AAEDB
aaedb_connect()

# process a basic query on the database
query_fn <- function(x) {
  tbl(x, in_schema(sql("aquatic_data"), sql("site"))) \%>\%
    filter(waterbody == "Ovens River")
}
site_info <- fetch_query(query_fn)

}
