% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_data.R
\name{fetch_data}
\alias{fetch_data}
\alias{fetch_table}
\alias{fetch_query}
\alias{fetch_project}
\title{Query the AAEDB}
\usage{
fetch_table(x, schema = "aquatic_data", ...)

fetch_query(query, ...)

fetch_project(project_id, range = c(1900, 2099), schema = "aquatic_data", ...)
}
\arguments{
\item{x}{a character specifying an individual table in the AAEDB}

\item{schema}{schema in which \code{x} is found. Defaults to
\code{"aquatic_data"}}

\item{\dots}{additional arguments passed to function-valued
\code{query} arguments in \code{fetch_query} (ignored otherwise)}

\item{query}{a character specifying a SQL query or a function
that takes a single argument (the database connection) and
processes a dplyr-style series of operations}

\item{project_id}{an integer specifying an individual AAE project
(1 - Snags, 2 - VEFMAP, 4 - NFRC, 6 - Kiewa Ovens, 7 - SRA,
 8 - Southern Basins, 9 - Ovens Demo Reaches, 10 - King Parrot
 Creek Macquarie Perch, 11 - Lower Goulburn Projects, 12 - Hughes
 Creek Macquarie Perch, 13 - Seven Creeks Macquarie Perch,
 14 - Index of Estuarine Condition, 15 - LTIM Lower Goulburn,
 16 - IVT Broken Creek)}

\item{range}{a vector containing minimum and maximum years for
which data are required}
}
\description{
\code{fetch_table}, \code{fetch_query}, and
  \code{fetch_project} represent three ways to interact with the
  AAEDB. \code{fetch_table} provides access to prepared tables in
  the database, \code{fetch_query} allows users to compute custom
  queries, and \code{fetch_project} prepares and downloads data
  for an individual AAE project.

  All functions require credentials to access the AAEDB, as well
  as a appropriate VPN connection. If making multiple queries,
  it can be easier to connect once to the AAEDB rather than repeatedly
  connecting (and disconnecting). This is possible with the
  \code{aaedb_connect} function.

  \code{fetch_query} can be used to download anything
  you would download with \code{fetch_table} or \code{fetch_project}.
  The benefits of \code{fetch_table} and \code{fetch_project} are in
  providing access to prepared tables containing commonly used variables
  for analyses of fish data. The benefit of \code{fetch_query} is in
  allowing custom queries and allowing processing of data files to occur
  on the server prior to downloading the final data set. This may be
  especially useful when working with large spatial data sets.
}
\examples{
# connect to the AAEDB
aaedb_connect()

# download the VEFMAP database flat file
vefmap <- fetch_table("v_vefmap_only_flat_data")

# process a simple SQL query to list all projects with data from the
#   Ovens river
survey_info <- fetch_query(
  "SELECT waterbody, id_project
     FROM aquatic_data.site a LEFT JOIN aquatic_data.survey b
     ON a.id_site = b.id_site
     WHERE lower(waterbody) LIKE 'ovens\%'
     GROUP BY waterbody, id_project
     ORDER by waterbody, id_project"
)

# process a dplyr query to collect site information
query_fn <- function(x) {
  tbl(x, in_schema(sql("aquatic_data"), sql("site"))) \%>\%
    filter(waterbody == "Ovens River")
}
ovens_sites <- fetch_query(query_fn)

# and grab information for individual sites
ovens_data <- fetch_project(9)

# possibly for a subset of years
ovens_data <- fetch_project(9, range = c(2015, 2017))

# optional: disconnect from the AAEDB prior to ending the R session
# aaedb_disconnect()

}
